LIstViewAdapter.cs: when a new task is added, the whole list is now sorted
DialogNewTask.cs and TaskBase.cs: a limit of characters in a title was added, when this limit is reached, the focus on the task is automatically disabled

Task.cs: deleted a method which adds the task to the database. It was moved to ListViewAdapter instead.

ToDoBase.cs: the tasks are now sorted when the ToDo list is created